functions/dha-doctorSummaryAndQA/requirements.txt
functions-framework==3.*
google-genai
Flask==3.0.3
Flask-Cors==5.0.0
